tohoku/hazard: HazardOptions
